requirements.txt
textual>=0.60.0
python-chi>=0.18.0
pyyaml>=6.0
paramiko>=3.0
keyring>=24.0
keyrings.cryptfile>=1.3
python-blazarclient>=1.1.1
